+++ docs/changelog.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- removed `addRootDocument`/`removeRootDocument` APIs (deprecated since 2.11.0)
+- [Issue #56](https://github.com/gullerya/data-tier/issues/56) - internal implementation improvements
+- documentation updated
+- dependencies maintanance
 
 ## [3.1.6] - 2021-04-24
 ### Added